django/requirements.txt
Django==2.0.3
mysqlclient==1.3.12
wheel==0.31.0
gunicorn==19.7.1
psycopg2==2.7.5
google-api-python-client==1.5.5
oauth2client==4.0.0
oauthlib==2.0.0
djangorestframework==3.5.3
python-social-auth==0.3.6
social-auth-app-django==2.1.0    
social-auth-core==1.7.0
django-cron==0.5.1
django-common-helpers==0.9.2
sendgrid==5.4.1
python-http-client==3.1.0
Pillow==5.2.0
django-filter==2.0.0
djangorestframework==3.8.2 
